# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: PTutor API
  version: 1.0.0
  description: API documentation for the PTutor tutoring platform.
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
  /api/v1/complaints:
    post:
      tags: [Complaints]
      summary: Submit a complaint
      description: A student or tutor submits a complaint for one of their contracts. Evidence URLs may be included in the same request.
      operationId: createComplaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComplaintRequest'
      responses:
        '201':
          description: Complaint submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    get:
      tags: [Complaints]
      summary: List my complaints
      operationId: listMyComplaints
      responses:
        '200':
          description: Complaints submitted by the authenticated student or tutor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintListResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/complaints/{complaintId}:
    get:
      tags: [Complaints]
      summary: Get one of my complaints
      operationId: getMyComplaint
      parameters:
        - $ref: '#/components/parameters/ComplaintId'
      responses:
        '200':
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/complaints/{complaintId}/evidences:
    post:
      tags: [Complaints]
      summary: Add evidence to my complaint
      operationId: addComplaintEvidence
      parameters:
        - $ref: '#/components/parameters/ComplaintId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintEvidenceRequest'
      responses:
        '201':
          description: Evidence added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintEvidenceResponseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/v1/complaints/{complaintId}/evidences/{evidenceId}:
    delete:
      tags: [Complaints]
      summary: Delete evidence from my complaint
      operationId: deleteComplaintEvidence
      parameters:
        - $ref: '#/components/parameters/ComplaintId'
        - name: evidenceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Evidence deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintDeleteResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/v1/admin/complaints:
    get:
      tags: [Complaints]
      summary: List complaints for processing
      operationId: listComplaintsForManagement
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, IN_REVIEW, RESOLVED, REJECTED]
      responses:
        '200':
          description: Complaints ordered from newest to oldest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintListResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/admin/complaints/{complaintId}:
    get:
      tags: [Complaints]
      summary: Get a complaint for processing
      operationId: getComplaintForManagement
      parameters:
        - $ref: '#/components/parameters/ComplaintId'
      responses:
        '200':
          description: Complaint details including evidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/complaints/{complaintId}/status:
    patch:
      tags: [Complaints]
      summary: Process a complaint
      description: IN_REVIEW represents requesting more evidence or continuing investigation. RESOLVED and REJECTED require a resolution message.
      operationId: processComplaint
      parameters:
        - $ref: '#/components/parameters/ComplaintId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintResolutionRequest'
      responses:
        '200':
          description: Complaint status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ComplaintId:
      name: complaintId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ComplaintEvidenceRequest:
      type: object
      required: [fileUrl, fileType]
      properties:
        fileUrl:
          type: string
          maxLength: 500
        fileType:
          type: string
          maxLength: 100
    CreateComplaintRequest:
      type: object
      required: [contractId, title, content]
      properties:
        contractId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        content:
          type: string
          maxLength: 10000
        evidences:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/ComplaintEvidenceRequest'
    ComplaintResolutionRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [IN_REVIEW, RESOLVED, REJECTED]
        resolution:
          type: string
          maxLength: 10000
          description: Required for RESOLVED and REJECTED.
    ComplaintEvidenceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileUrl:
          type: string
        fileType:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ComplaintResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [PENDING, IN_REVIEW, RESOLVED, REJECTED]
        resolution:
          type: string
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        evidences:
          type: array
          items:
            $ref: '#/components/schemas/ComplaintEvidenceResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiResponseMetadata:
      type: object
      required:
        - success
        - code
        - message
        - errors
        - timestamp
        - path
      properties:
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
...
    ComplaintResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseMetadata'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ComplaintResponse'
    ComplaintListResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseMetadata'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ComplaintResponse'
    ComplaintEvidenceResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseMetadata'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ComplaintEvidenceResponse'
    ComplaintDeleteResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseMetadata'
        - type: object
          properties:
            data:
              nullable: true
              example: null
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseMetadata'
        - type: object
          properties:
            data:
              nullable: true
              example: null
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: VALIDATION_ERROR
            message: Request validation failed
            data: null
            errors:
              email: Invalid email
            timestamp: '2026-08-30T10:00:00Z'
            path: /api/v1/auth/register
    UnauthorizedError:
      description: Authentication failed or token is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: INVALID_CREDENTIALS
            message: Email or password is incorrect
            data: null
            errors: {}
            timestamp: '2026-08-30T10:00:00Z'
            path: /api/v1/auth/login
    ForbiddenError:
      description: The authenticated user does not have permission to access the requested resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: FORBIDDEN
            message: You do not have permission to access this resource
            data: null
            errors: {}
            timestamp: '2026-08-30T10:00:00Z'
            path: /api/v1/students/me
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: STUDENT_PROFILE_NOT_FOUND
            message: Student profile not found
            data: null
            errors: {}
            timestamp: '2026-08-30T10:00:00Z'
            path: /api/v1/students/me
...
    ConflictError:
      description: Request conflicts with existing data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: EMAIL_ALREADY_EXISTS
            message: Email is already registered
            data: null
            errors: {}
            timestamp: '2026-08-30T10:00:00Z'
            path: /api/v1/auth/register
...